Cast fix and moving Win32 stuff in a dedicated file.

+++ Form1.cs
@@ -132,7 +132,7 @@ namespace RemoteControlSample
                 //Display MCE button name
                 label1.Text = e.MceButton.ToString();
                 //Check if this is an HP extension
-                if (Enum.IsDefined(typeof(Hid.UsageTables.HpMceButton), (int)e.MceButton))
+                if (Enum.IsDefined(typeof(Hid.UsageTables.HpMceButton), (ushort)e.MceButton))
                 {
                     //Also display HP button name
                     label1.Text += " / HP:" + ((Hid.UsageTables.HpMceButton)e.MceButton).ToString();
